gh-156570: move warnings helper regression test
